feat: Migrate from .config/.tfvars for module
As part of a general migration of this repo to be fully declarative, the
root module is rewritten to be unaware of the backend used for state.

Fixed some dependency issues found when recreating the environments; add
`depends_on` as needed.
Removed the changes to default Compute and App Engine service accounts.